requirements.txt
numpy==1.19.2
streamlit==1.0.0
tensorflow_hub==0.12.0
tensorflow==2.4.0
Pillow==8.4.0